// src/store/action-types.ts
/* eslint-disable */
export enum ActionType {
  AUTH_FAIL = 'AUTHENTICATION_FAILURE',
  AUTH_REQUIRED = 'AUTHORIZATION_REQUIRED',
  CHANGE_GENRE = 'CHANGE_GENRE',
  GET_NETWORK_ERROR = 'GET_NETWORK_ERROR',
  GET_PROMO_MOVIE = 'GET_PROMO_MOVIE',
  GET_REVIEWS = 'GET_REVIEWS',
  GET_USER_DATA = 'GET_USER_DATA',
  LOAD_MOVIES = 'LOAD_MOVIES'
}